Revert "creation of new entity Topic and relations with student and mentor entities"
This reverts commit 4c2bb8194a76a820326fbed4c7d1b4880e3ac2a5.
