+++ src/Entity/Mentor.php
@@ -50,15 +50,28 @@ class Mentor
     private ?User $user;
 
     /**
+     * @ORM\Column(type="integer")
+     * @Assert\NotBlank(
+     *      message = "Merci de choisir au moins un sujet de mentorat.")
+     */
+    private int $topic1;
+
+    /**
+     * @ORM\Column(type="integer", nullable=true)
+     * @Assert\Choice ({1,2,3,4,5,6,7,8,9, null})
+     */
+    private ?int $topic2;
+
+    /**
+     * @ORM\Column(type="integer", nullable=true)
+     * @Assert\Choice ({1,2,3,4,5,6,7,8,9, null})
+     */
+    private ?int $topic3;
+
+    /**
      * @ORM\OneToOne(targetEntity=Mentoring::class, inversedBy="mentor", cascade={"persist", "remove"})
      */
     private ?Mentoring $mentoring;
-
-    /**
-     * @ORM\ManyToOne(targetEntity=Topic::class, inversedBy="mentors", cascade={"persist", "remove"})
-     * @ORM\JoinColumn(nullable=false)
-     */
-    private Topic $topics;
 
     public function getId(): ?int
     {
@@ -113,6 +126,42 @@ class Mentor
         return $this;
     }
 
+    public function getTopic1(): ?int
+    {
+        return $this->topic1;
+    }
+
+    public function setTopic1(int $topic1): self
+    {
+        $this->topic1 = $topic1;
+
+        return $this;
+    }
+
+    public function getTopic2(): ?int
+    {
+        return $this->topic2;
+    }
+
+    public function setTopic2(?int $topic2): self
+    {
+        $this->topic2 = $topic2;
+
+        return $this;
+    }
+
+    public function getTopic3(): ?int
+    {
+        return $this->topic3;
+    }
+
+    public function setTopic3(?int $topic3): self
+    {
+        $this->topic3 = $topic3;
+
+        return $this;
+    }
+
     public function getUser(): ?User
     {
         return $this->user;
@@ -133,18 +182,6 @@ class Mentor
     public function setMentoring(?Mentoring $mentoring): self
     {
         $this->mentoring = $mentoring;
-
-        return $this;
-    }
-
-    public function getTopics(): Topic
-    {
-        return $this->topics;
-    }
-
-    public function setTopics(Topic $topics): self
-    {
-        $this->topics = $topics;
 
         return $this;
     }
